pyproject.toml
[tool.poetry]
name = "leksuss_gendiff"
version = "0.3.2"
description = ""
authors = ["Andrey <leksus@leksus.info>"]
repository = "https://github.com/leksuss/python-project-lvl2"
packages = [
    { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.13.4"
asciinema = "^2.0.2"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
